fix(claude): persist worktrees, fix resume, refactor auth
- Move worktrees to ~/.claude-api/worktrees (persistent on PVC)
- Fix conversation resume: use --resume without -p flag
- Extract ttyd auth logic into separate auth.ts module
- Ensure workdir exists before spawning Claude
