+++ charts/claude/src/src/index.ts
@@ -1,12 +1,13 @@
-import express, { Request, Response } from "express";
+import express from "express";
 import { WebSocketServer, WebSocket } from "ws";
-import { createServer, ServerResponse, IncomingMessage } from "http";
+import { createServer, IncomingMessage } from "http";
 import { Socket } from "net";
-import { createProxyMiddleware } from "http-proxy-middleware";
 import { v4 as uuidv4 } from "uuid";
 import { spawn, ChildProcess } from "child_process";
 import path from "path";
 import fs from "fs";
+
+import { setupAuthRoutes, handleTtydUpgrade } from "./auth";
 
 const app = express();
 app.use(express.json());
@@ -17,14 +18,10 @@ const WORKTREES_DIR = path.join(HOME, ".claude-api", "worktrees"); // Persistent
 const SESSIONS_DIR = path.join(HOME, ".claude-api", "sessions");
 const STATIC_DIR = process.env.STATIC_DIR || "/app/public";
 const CLAUDE_BIN = path.join(HOME, ".npm-global", "bin", "claude");
-const TTYD_PORT = 7681;
 
 // Ensure directories exist
 fs.mkdirSync(SESSIONS_DIR, { recursive: true });
 fs.mkdirSync(WORKTREES_DIR, { recursive: true });
-
-// Authentication state
-let authTtydProcess: ChildProcess | null = null;
 
 interface Session {
   id: string;
@@ -44,117 +41,8 @@ app.get("/api/health", (_req, res) => {
   res.json({ status: "ok", sessions: sessions.size });
 });
 
-// Auth status - check if Claude is authenticated
-app.get("/api/auth/status", (_req, res) => {
-  const authFile = path.join(HOME, ".claude", "auth.json");
-  const authenticated = fs.existsSync(authFile);
-
-  res.json({
-    authenticated,
-    terminalActive: authTtydProcess !== null,
-  });
-});
-
-// Start auth terminal (spawn ttyd)
-app.post("/api/auth/start", (_req, res) => {
-  // Clean up any existing auth process
-  if (authTtydProcess) {
-    console.log("Killing existing ttyd process...");
-    authTtydProcess.kill();
-    authTtydProcess = null;
-  }
-
-  try {
-    // Spawn ttyd with claude /login
-    // -W: Don't wait for initial connection (start immediately)
-    // -p: Port to listen on
-    // -t: Set terminal type
-    console.log(`Starting ttyd on port ${TTYD_PORT}...`);
-
-    // Run claude interactively - user types /login inside the session
-    const ttyd = spawn(
-      "ttyd",
-      [
-        "-p",
-        TTYD_PORT.toString(),
-        "-W", // Start immediately
-        "-t",
-        "titleFixed=Claude Authentication",
-        CLAUDE_BIN,
-      ],
-      {
-        cwd: HOME,
-        env: { ...process.env, HOME },
-        stdio: ["ignore", "pipe", "pipe"],
-      },
-    );
-
-    authTtydProcess = ttyd;
-
-    ttyd.stdout.on("data", (data) => {
-      console.log(`[ttyd stdout] ${data.toString().trim()}`);
-    });
-
-    ttyd.stderr.on("data", (data) => {
-      console.log(`[ttyd stderr] ${data.toString().trim()}`);
-    });
-
-    ttyd.on("close", (code) => {
-      console.log(`ttyd process exited with code ${code}`);
-      authTtydProcess = null;
-    });
-
-    ttyd.on("error", (err) => {
-      console.error(`ttyd process error: ${err}`);
-      authTtydProcess = null;
-    });
-
-    // Give ttyd a moment to start
-    setTimeout(() => {
-      res.json({
-        success: true,
-        message: "Terminal started. Connect to /api/auth/terminal",
-        terminalUrl: "/api/auth/terminal",
-      });
-    }, 500);
-  } catch (err) {
-    console.error("Failed to start ttyd:", err);
-    res.status(500).json({ error: "Failed to start terminal" });
-  }
-});
-
-// Stop auth terminal
-app.post("/api/auth/stop", (_req, res) => {
-  if (authTtydProcess) {
-    console.log("Stopping ttyd process...");
-    authTtydProcess.kill();
-    authTtydProcess = null;
-  }
-  res.json({ success: true });
-});
-
-// Proxy to ttyd terminal (HTTP only - WebSocket handled separately)
-const ttydProxy = createProxyMiddleware({
-  target: `http://localhost:${TTYD_PORT}`,
-  changeOrigin: true,
-  pathRewrite: {
-    "^/api/auth/terminal": "", // Remove prefix
-  },
-  on: {
-    error: (
-      err: Error,
-      req: Request,
-      res: Response | ServerResponse | Socket,
-    ) => {
-      console.error("Proxy error:", err);
-      if (res instanceof ServerResponse) {
-        res.writeHead(502);
-        res.end("Terminal not available");
-      }
-    },
-  },
-});
-app.use("/api/auth/terminal", ttydProxy);
+// Setup auth routes (ttyd terminal for /login)
+setupAuthRoutes(app);
 
 // List sessions
 app.get("/api/sessions", (_req, res) => {
@@ -190,7 +78,7 @@ app.post("/api/sessions", (req, res) => {
 
   sessions.set(id, session);
   console.log(
-    `Session created: ${id}, Total sessions in memory: ${sessions.size}`,
+    `Session created: ${id}, Total sessions in memory: ${sessions.size}`
   );
 
   // Save session metadata
@@ -202,7 +90,7 @@ app.post("/api/sessions", (req, res) => {
       name: session.name,
       workdir: session.workdir,
       createdAt: session.createdAt,
-    }),
+    })
   );
 
   res.status(201).json({
@@ -270,103 +158,16 @@ app.get("*", (req, res) => {
 // Create HTTP server
 const server = createServer(app);
 
-// WebSocket server for ttyd terminal proxy
-// Uses proper WebSocket-to-WebSocket proxying (like ttyd-session-manager did with gorilla/websocket)
-// IMPORTANT: Must accept "tty" subprotocol or ttyd client will reject the connection
-// IMPORTANT: Disable compression on server - browser may negotiate it but ttyd doesn't use it
-const ttydWss = new WebSocketServer({
-  noServer: true,
-  perMessageDeflate: false,
-  handleProtocols: (protocols) => {
-    // Accept "tty" subprotocol if client requests it (ttyd always does)
-    if (protocols.has("tty")) {
-      return "tty";
-    }
-    return false;
-  },
-});
-
-ttydWss.on("connection", (clientWs: WebSocket) => {
-  console.log("Client WebSocket connected, connecting to ttyd...");
-
-  // Connect to ttyd using WebSocket protocol with "tty" subprotocol
-  // IMPORTANT: Disable perMessageDeflate to avoid compression mismatch with ttyd
-  const ttydWs = new WebSocket(`ws://localhost:${TTYD_PORT}/ws`, ["tty"], {
-    perMessageDeflate: false,
-  });
-
-  // Set binary type to match ttyd expectations
-  ttydWs.binaryType = "arraybuffer";
-
-  // Track connection state
-  let ttydConnected = false;
-
-  ttydWs.on("open", () => {
-    console.log("Connected to ttyd WebSocket");
-    ttydConnected = true;
-  });
-
-  ttydWs.on("message", (data: Buffer, isBinary: boolean) => {
-    // Forward ttyd messages to client, preserving binary/text type
-    if (clientWs.readyState === WebSocket.OPEN) {
-      clientWs.send(data, { binary: isBinary });
-    }
-  });
-
-  ttydWs.on("close", (code, reason) => {
-    console.log(`ttyd WebSocket closed: ${code} ${reason}`);
-    ttydConnected = false;
-    if (clientWs.readyState === WebSocket.OPEN) {
-      clientWs.close(code, reason.toString());
-    }
-  });
-
-  ttydWs.on("error", (err) => {
-    console.error("ttyd WebSocket error:", err);
-    ttydConnected = false;
-    if (clientWs.readyState === WebSocket.OPEN) {
-      clientWs.close(1011, "ttyd connection error");
-    }
-  });
-
-  clientWs.on("message", (data: Buffer, isBinary: boolean) => {
-    // Forward client messages to ttyd, preserving binary/text type
-    if (ttydWs.readyState === WebSocket.OPEN) {
-      ttydWs.send(data, { binary: isBinary });
-    }
-  });
-
-  clientWs.on("close", (code, reason) => {
-    console.log(`Client WebSocket closed: ${code} ${reason}`);
-    // Only close ttyd connection if it's open or connecting
-    if (ttydConnected || ttydWs.readyState === WebSocket.CONNECTING) {
-      ttydWs.close();
-    }
-  });
-
-  clientWs.on("error", (err) => {
-    console.error("Client WebSocket error:", err);
-    if (ttydConnected || ttydWs.readyState === WebSocket.CONNECTING) {
-      ttydWs.close();
-    }
-  });
-});
-
-// WebSocket server for streaming (also noServer to avoid duplicate upgrade handlers)
+// WebSocket server for session streaming
 const wss = new WebSocketServer({ noServer: true });
 
 // Handle ALL WebSocket upgrades in one place to avoid conflicts
-// When using { server, path } option, ws library registers its own upgrade handler
-// which can conflict with manual handlers and send duplicate responses
 server.on("upgrade", (req: IncomingMessage, socket: Socket, head: Buffer) => {
   const url = req.url || "";
-  console.log(`[UPGRADE] Request received: ${url}, headers: ${JSON.stringify(req.headers)}`);
+  console.log(`[UPGRADE] Request received: ${url}`);
 
   if (url.startsWith("/api/auth/terminal/ws")) {
-    console.log(`WebSocket upgrade request for ttyd: ${url}`);
-    ttydWss.handleUpgrade(req, socket, head, (ws) => {
-      ttydWss.emit("connection", ws, req);
-    });
+    handleTtydUpgrade(req, socket, head);
   } else if (url.startsWith("/ws")) {
     console.log(`WebSocket upgrade request for session: ${url}`);
     wss.handleUpgrade(req, socket, head, (ws) => {
@@ -402,7 +203,7 @@ wss.on("connection", (ws, req) => {
   const session = sessions.get(sessionId);
   if (!session) {
     console.log(
-      `Session ${sessionId} not found. Available sessions: ${Array.from(sessions.keys()).join(", ")}`,
+      `Session ${sessionId} not found. Available sessions: ${Array.from(sessions.keys()).join(", ")}`
     );
     ws.close(4004, "Session not found");
     return;
@@ -416,28 +217,34 @@ wss.on("connection", (ws, req) => {
   ws.on("message", (data) => {
     const message = JSON.parse(data.toString());
     console.log(
-      `Received message for session ${session.id}: ${JSON.stringify(message).substring(0, 100)}`,
+      `Received message for session ${session.id}: ${JSON.stringify(message).substring(0, 100)}`
     );
 
     if (message.type === "input") {
       if (session.isProcessing) {
-        console.log(`Session ${session.id} is already processing, queuing not implemented`);
-        ws.send(JSON.stringify({
-          type: "error",
-          content: "Please wait for the current response to complete",
-        }));
+        console.log(
+          `Session ${session.id} is already processing, queuing not implemented`
+        );
+        ws.send(
+          JSON.stringify({
+            type: "error",
+            content: "Please wait for the current response to complete",
+          })
+        );
         return;
       }
 
       // Run Claude in print mode with the user's message
-      console.log(`Running Claude for session ${session.id} with message: ${message.content.substring(0, 50)}...`);
+      console.log(
+        `Running Claude for session ${session.id} with message: ${message.content.substring(0, 50)}...`
+      );
       runClaudeMessage(session, message.content);
     }
   });
 
   ws.on("close", (code, reason) => {
     console.log(
-      `Session ${session.id} WebSocket closed: ${code} ${reason.toString()}`,
+      `Session ${session.id} WebSocket closed: ${code} ${reason.toString()}`
     );
     session.wsClients.delete(ws);
   });
@@ -455,18 +262,21 @@ wss.on("connection", (ws, req) => {
         sessionId: session.id,
         name: session.name,
         workdir: session.workdir,
-      }),
+      })
     );
     console.log(`Welcome message sent successfully to session ${session.id}`);
   } catch (err) {
-    console.error(`Failed to send welcome message to session ${session.id}:`, err);
+    console.error(
+      `Failed to send welcome message to session ${session.id}:`,
+      err
+    );
   }
 });
 
 // Run Claude in print mode with a single message (like cui does)
 function runClaudeMessage(session: Session, userMessage: string) {
   console.log(
-    `Running Claude for session ${session.id} in ${session.workdir}`,
+    `Running Claude for session ${session.id} in ${session.workdir}`
   );
   console.log(`Using Claude binary: ${CLAUDE_BIN}`);
 
@@ -478,25 +288,26 @@ function runClaudeMessage(session: Session, userMessage: string) {
 
   session.isProcessing = true;
 
-  // Build args for print mode
-  // -p: print mode (non-interactive)
-  // --output-format stream-json: structured JSONL output
-  // --verbose: required with stream-json
-  // --dangerously-skip-permissions: skip permission prompts
-  const args = [
-    "-p", // Print mode
-    "--output-format", "stream-json",
-    "--verbose",
-    "--dangerously-skip-permissions",
-  ];
+  // Build args based on whether this is a new conversation or resume
+  // cui pattern:
+  //   new: -p <message> --output-format stream-json --verbose
+  //   resume: --resume <sessionId> <message> --output-format stream-json --verbose
+  const args: string[] = [];
 
-  // If we have a previous Claude session, resume it
   if (session.claudeSessionId) {
+    // Resume existing conversation (no -p flag)
     args.push("--resume", session.claudeSessionId);
+    args.push(userMessage);
+  } else {
+    // New conversation with print mode
+    args.push("-p");
+    args.push(userMessage);
   }
 
-  // Add the user message as the last argument
-  args.push(userMessage);
+  // Common flags
+  args.push("--output-format", "stream-json");
+  args.push("--verbose");
+  args.push("--dangerously-skip-permissions");
 
   console.log(`Claude args: ${args.join(" ")}`);
 
@@ -519,7 +330,10 @@ function runClaudeMessage(session: Session, userMessage: string) {
     console.error(`Claude process error: ${err.message}`);
     session.isProcessing = false;
     session.process = undefined;
-    broadcast(session, { type: "error", content: `Process error: ${err.message}` });
+    broadcast(session, {
+      type: "error",
+      content: `Process error: ${err.message}`,
+    });
   });
 
   // Buffer for incomplete JSON lines
@@ -539,7 +353,7 @@ function runClaudeMessage(session: Session, userMessage: string) {
       try {
         const event = JSON.parse(line);
         handleClaudeEvent(session, event);
-      } catch (err) {
+      } catch {
         // Not valid JSON, might be regular output
         console.log(`Claude non-JSON output: ${line.substring(0, 100)}`);
       }
@@ -589,7 +403,9 @@ function handleClaudeEvent(session: Session, event: Record<string, unknown>) {
 
   // Forward assistant messages to clients
   if (event.type === "assistant") {
-    const msg = event as { message?: { content?: Array<{ type: string; text?: string }> } };
+    const msg = event as {
+      message?: { content?: Array<{ type: string; text?: string }> };
+    };
     if (msg.message?.content) {
       for (const block of msg.message.content) {
         if (block.type === "text" && block.text) {
@@ -610,7 +426,10 @@ function handleClaudeEvent(session: Session, event: Record<string, unknown>) {
   // Handle errors
   if (event.type === "error") {
     const err = event as { error?: { message?: string } };
-    broadcast(session, { type: "error", content: err.error?.message || "Unknown error" });
+    broadcast(session, {
+      type: "error",
+      content: err.error?.message || "Unknown error",
+    });
   }
 }
 
@@ -635,7 +454,7 @@ function saveSession(session: Session) {
       workdir: session.workdir,
       createdAt: session.createdAt,
       claudeSessionId: session.claudeSessionId,
-    }),
+    })
   );
 }
 
@@ -649,7 +468,7 @@ function loadSessions() {
 
     try {
       const data = JSON.parse(
-        fs.readFileSync(path.join(SESSIONS_DIR, file), "utf-8"),
+        fs.readFileSync(path.join(SESSIONS_DIR, file), "utf-8")
       );
       sessions.set(data.id, {
         id: data.id,
